Updated for product and admin（regulated api)

+++ src/main/java/sg/nusiss/t6/caproject/repository/ReviewRepository.java
@@ -7,12 +7,13 @@ import sg.nusiss.t6.caproject.model.Review;
 import java.util.List;
 
 @Repository
-public interface ReviewRepository extends JpaRepository<Review, Long> {
+public interface ReviewRepository extends JpaRepository<Review, Integer> {
 
     /**
      * 根据商品ID查找所有评论
+     * 
      * @param productId 商品ID
      * @return 评论列表
      */
-    List<Review> findByProductId(Long productId);
+    List<Review> findByProductId(Integer productId);
 }
